@@ -1,8 +1,7 @@
 import { Component } from '@angular/core';
 import {FormControl, FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {TooltipPosition, MatTooltipModule} from '@angular/material/tooltip';
+import {MatTooltipModule} from '@angular/material/tooltip';
 import {MatButtonModule} from '@angular/material/button';
-import {CdkScrollable} from '@angular/cdk/scrolling';
 import {MatSelectModule} from '@angular/material/select';
 import {MatFormFieldModule} from '@angular/material/form-field';
 import { MatCardModule } from '@angular/material/card';
@@ -17,7 +16,6 @@ import { MatCheckboxModule } from '@angular/material/checkbox';
     MatSelectModule,
     FormsModule,
     ReactiveFormsModule,
-    CdkScrollable,
     MatButtonModule,
     MatTooltipModule, MatCardModule, MatInputModule, MatCheckboxModule
   ],
